refactor(server): merge safety_cache.py into safety.py
The cache (`SAFETY_CACHE_FILE`, `load_safety_cache`, `save_safety_cache`,
the TypeAdapter) only exists to back the NSFW classifier — same area
of responsibility, so collapse the two files. `safety.py` now owns
both the classifier and its disk-backed result cache.

Lazy-load preservation: importing safety.py pulls torch / timm /
transformers transitively, so consumers that only need the cache
(ws_session.py's handler bodies) use a function-level import to avoid
dragging the heavy import waterfall into ws_session's module graph.
server.py's heavy try-block now imports `load_safety_cache` alongside
`SafetyChecker`, keeping torch's load instrumented.

Refs #78.
